fake_drivers/requirements.txt
Pyrebase==3.0.27
python-dotenv==0.6.3
